+++ CHANGELOG.md
@@ -24,6 +24,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `ExplicitTObjectRule` analysis rule, which flags `class` declarations that don't explicitly
   declare their parent type.
 - `MethodNestingDepthRule` analysis rule, which flags subroutines that are nested too deeply.
+- `SpecialKeywordCapitalizationRule` analysis rule, which flags special keywords that are not
+  capitalized.
 - `ForbiddenIdentifierRule` analysis rule template, which flags identifiers with a specified name.
 - `ForbiddenMethodRule` analysis rule template, which flags methods with a specified fully-qualified
   name.